chore(benchmark): ordered utilitary methods in utils subnamespace
