+++ tests/benchmark/utils/__init__.py
@@ -1,0 +1,7 @@
+"""Utilitary methods for benchmarking processes."""
+
+from .infos import extract_nsteps
+from .structures import BenchmarkEntry, SEProblem
+from .unpacking import maybe_unpack
+
+__all__ = ['BenchmarkEntry', 'SEProblem', 'extract_nsteps', 'maybe_unpack']
